Implement and test templater
tampl is a small application that expects two arguments, first the path
to the source folder which contains a `_vars.yaml` and a number of
`*.tmpl` files, and second the target or output path where the rendered
templates are saved.

A target's filename is simply the template's filename /wo the `.tmpl`
extension.
Variables are defined in `_vars.yml` and are available in all templates.