Simpli- and prettify template
Note that the test is not deterministic at the moment because the order
of keys in a go dictionary is not specified. Either I have to sort the
keys in the template, which is quite tedious or I have to use a single
host in the SSH config example.
I tend to choose the latter option.